+++ clang-tools-extra/test/clang-tidy/checkers/misc/const-correctness-pointer-as-values.cpp
@@ -33,7 +33,9 @@ void range_for() {
   int *p_local2[2] = {nullptr, nullptr};
   // CHECK-MESSAGES: [[@LINE-1]]:3: warning: variable 'p_local2' of type 'int *[2]' can be declared 'const'
   // CHECK-FIXES: int *const p_local2[2]
-  for (const auto *con_ptr : p_local2) {
+  for (const auto *p_local3 : p_local2) {
+  // CHECK-MESSAGES: [[@LINE-1]]:8: warning: variable 'p_local3' of type 'const int *' can be declared 'const'
+  // CHECK-FIXES: for (const auto *const p_local3 : p_local2)
   }
 
 }
@@ -62,6 +64,8 @@ void EmitProtocolMethodList(T &&Methods) {
   // CHECK-FIXES: SmallVector<const int *> const p_local0
   SmallVector<const int *> np_local0;
   for (const auto *I : Methods) {
+  // CHECK-MESSAGES: [[@LINE-1]]:8: warning: variable 'I' of type 'const int *' can be declared 'const'
+  // CHECK-FIXES: for (const auto *const I : Methods)
     if (I == nullptr)
       np_local0.push_back(I);
   }